# openapi.yaml
openapi: 3.0.0
info:
  title: QoC API
  description: Documentation Swagger pour une API exemple.
  version: 0.0.1
servers:
  - url: http://localhost:3000
paths:
...
  /users:
    get:
      summary: Get users
      tags:
        - Users
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
    post:
      summary: Create a user
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: Ok 
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
...
components:
  schemas:
...
    User:
      type: object
      required:
        - age
        - sex
        - job
        - politicalOpinion
        - nationality
      properties:
        id:
          type: integer
        age:
          type: integer
        sex:
          type: string
        job:
          type: string
        politicalOpinion:
          type: string
        nationality:
          type: string
...